requirements.txt
requests
pytest
pytest-rerunfailures
pytest-testconfig
pytest-logger
jsonschema
urllib3
pytruth
paho-mqtt
colorlog
allure-pytest
